Move minimum_version to .util.importlib
Allows to import in other .util submodules.
